requirements.txt
transformers>=4.51.0
torch>=2.7.0
qdrant-client>=1.17.0
z3-solver>=4.16.0
sentence-transformers>=5.4.0
python-dotenv>=1.1.0
accelerate>=1.6.0
bitsandbytes>=0.49.0
gradio>=5.0.0
peft>=0.15.0
trl>=0.18.0
datasets>=3.5.0
pytest>=9.0.0
